requirements.txt
requests>=2.25.0
aiohttp>=3.8.0
python-telegram-bot>=20.0 